Moved EnsembleTracker and did some refactoring
